+++ CRM/Nihrbackbone/BackboneConfig.php
@@ -45,6 +45,8 @@ class CRM_Nihrbackbone_BackboneConfig {
   private $_volunteerGeneralObservationsCustomGroup = [];
   private $_volunteerLifestyleCustomGroup = [];
   private $_volunteerSelectionEligibilityCustomGroup = [];
+  private $_volunteerLifeQualityCustomGroup = [];
+  private $_participationInStudiesCustomGroup = [];
   private $_selectionCriteriaCustomGroup = [];
   private $_volunteerIdsCustomGroup = [];
   private $_volunteerPanelCustomGroup = [];
@@ -299,6 +301,24 @@ class CRM_Nihrbackbone_BackboneConfig {
     }
   }
 
+  public function getVolunteerLifeQualityCustomGroup($key = NULL) {
+    if ($key && isset($this->_volunteerLifeQualityCustomGroup[$key])) {
+      return $this->_volunteerLifeQualityCustomGroup[$key];
+    }
+    else {
+      return $this->_volunteerLifeQualityCustomGroup;
+    }
+  }
+
+  public function getParticipationInStudiesCustomGroup($key = NULL) {
+    if ($key && isset($this->_participationInStudiesCustomGroup[$key])) {
+      return $this->_participationInStudiesCustomGroup[$key];
+    }
+    else {
+      return $this->_participationInStudiesCustomGroup;
+    }
+  }
+
   /**
    * Getter for volunteer panel custom group
    *
@@ -537,6 +557,35 @@ class CRM_Nihrbackbone_BackboneConfig {
     }
     return FALSE;
   }
+
+  public function getVolunteerLifeQualityCustomField($customFieldName, $key = NULL) {
+    foreach ($this->_volunteerLifeQualityCustomGroup['custom_fields'] as $customField) {
+      if ($customField['name'] == $customFieldName) {
+        if ($key && isset($customField[$key])) {
+          return $customField[$key];
+        }
+        else {
+          return $customField;
+        }
+      }
+    }
+    return FALSE;
+  }
+
+  public function getParticipationInStudiesCustomField($customFieldName, $key = NULL) {
+    foreach ($this->_participationInStudiesCustomGroup['custom_fields'] as $customField) {
+      if ($customField['name'] == $customFieldName) {
+        if ($key && isset($customField[$key])) {
+          return $customField[$key];
+        }
+        else {
+          return $customField;
+        }
+      }
+    }
+    return FALSE;
+  }
+
 
   /**
    * Getter for site alias data custom field
@@ -805,6 +854,12 @@ class CRM_Nihrbackbone_BackboneConfig {
       'gender',
       'nbr_consent_version',
       'case_status',
+      'nbr_employment_status',
+      'nbr_life_quality_general',
+      'nbr_life_quality_happiness',
+      'nbr_life_quality_overall',
+      'nbr_type_of_study',
+      'nbr_yes_no_dontknow'
     ];
     try {
       $foundOptionGroups = civicrm_api3('OptionGroup', 'get', [
@@ -1014,6 +1069,8 @@ class CRM_Nihrbackbone_BackboneConfig {
       'nihr_volunteer_status',
       'nihr_volunteer_general_observations',
       'nihr_volunteer_lifestyle',
+      'nihr_volunteer_life_quality',
+      'nihr_participation_in_studies',
       'nihr_volunteer_selection_eligibility',
       'nbr_selection_criteria',
       'nihr_volunteer_alias',