+++ CRM/Nihrbackbone/Form/NbrStudy.php
@@ -38,6 +38,11 @@ class CRM_Nihrbackbone_Form_NbrStudy extends CRM_Core_Form {
         ['type' => 'cancel', 'name' => E::ts('Done'), 'isDefault' => TRUE],
       ]);
     }
+    // check the study clone information if relevant
+    if (class_exists('CRM_Nbrclonestudy_BAO_NbrStudyClone')) {
+      $this->assign('clone_of', CRM_Nbrclonestudy_BAO_NbrStudyClone::setCloneOfStudyForm((int) $this->_studyId));
+      $this->assign('has_clones', CRM_Nbrclonestudy_BAO_NbrStudyClone::setHasClonesStudyForm((int) $this->_studyId));
+    }
     // export form elements
     $this->assign('elementNames', $this->getRenderableElementNames());
     parent::buildQuickForm();
@@ -54,10 +59,11 @@ class CRM_Nihrbackbone_Form_NbrStudy extends CRM_Core_Form {
       ]);
     $this->add('datepicker', 'start_date', E::ts('Start Date'), [],TRUE, ['time' => FALSE]);
     $this->add('datepicker', 'end_date', E::ts('End Date'), [],FALSE, ['time' => FALSE]);
-    $this->addEntityRef('nsd_researcher', E::ts('Researcher'), [
+    $this->addEntityRef('researcher_id', E::ts('Researcher(s)'), [
+      'multiple' => TRUE,
+      'placeholder' => '- select researcher(s)',
       'api' => ['params' => ['contact_sub_type' => 'nihr_researcher']],
-      'placeholder' => '- select researcher -',
-    ], FALSE);
+      ]);
     $this->addEntityRef('nsd_centre_origin', E::ts('Centre of Origin'), [
       'api' => ['params' => ['contact_sub_type' => 'nbr_centre']],
       'placeholder' => '- select centre -',
@@ -78,6 +84,7 @@ class CRM_Nihrbackbone_Form_NbrStudy extends CRM_Core_Form {
     $this->add('wysiwyg', 'nsd_scientific_info', E::ts('Scientific Information'), ['rows' => 4, 'cols' => 100], FALSE);
     $this->add('text', 'nsd_lay_title', E::ts('Lay title'), ['size' => 100], FALSE);
     $this->add('wysiwyg', 'nsd_lay_summary', E::ts('Lay summary'), ['rows' => 4, 'cols' => 100], FALSE);
+    $this->add('wysiwyg', 'nsd_study_outcome', E::ts('Study outcome'), ['rows' => 4, 'cols' => 100], FALSE);
     $this->add('text', 'nsd_study_long_name', E::ts("Long Name"), ['size' => 100], FALSE);
     $this->add('text', 'nsd_ethics_number', E::ts("Ethics Number"), [], FALSE);
     $this->add('advcheckbox', 'nsd_ethics_approved', E::ts('Ethics approved?'), [], FALSE);
@@ -110,8 +117,8 @@ class CRM_Nihrbackbone_Form_NbrStudy extends CRM_Core_Form {
     $this->add('text', 'nsc_age_to', E::ts("Age to"), [], FALSE);
     $this->add('text', 'nsc_bmi_from', E::ts("BMI from"), [], FALSE);
     $this->add('text', 'nsc_bmi_to', E::ts("BMI to"), [], FALSE);
-
   }
+
   /**
    * Add elements for view action
    */
@@ -124,11 +131,12 @@ class CRM_Nihrbackbone_Form_NbrStudy extends CRM_Core_Form {
     ]);
     $this->add('datepicker', 'start_date', E::ts('Start Date'), ['disabled' => 'disabled'],TRUE, ['time' => FALSE]);
     $this->add('datepicker', 'end_date', E::ts('End Date'), ['disabled' => 'disabled'],FALSE, ['time' => FALSE]);
-    $this->addEntityRef('nsd_researcher', E::ts('Researcher'), [
-      'api' => ['params' => ['contact_sub_type' => 'nihr_researcher']],
-      'placeholder' => '- select researcher -',
+    $this->addEntityRef('researcher_id', E::ts('Researcher(s)'), [
+      'multiple' => TRUE,
+      'placeholder' => '- select researcher(s)',
       'disabled' => 'disabled',
-    ], FALSE);
+      'api' => ['params' => ['contact_sub_type' => 'nihr_researcher']],
+    ]);
     $this->add('textarea', 'nsd_scientific_info', E::ts('Scientific Information'), [
       'rows' => 4,
       'cols' => 100,
@@ -136,6 +144,11 @@ class CRM_Nihrbackbone_Form_NbrStudy extends CRM_Core_Form {
       ], FALSE);
     $this->add('text', 'nsd_lay_title', E::ts('Lay title'), ['size' => 100,'disabled' => 'disabled'], FALSE);
     $this->add('textarea', 'nsd_lay_summary', E::ts('Lay summary'), [
+      'rows' => 4,
+      'cols' => 100,
+      'disabled' => 'disabled',
+    ], FALSE);
+    $this->add('textarea', 'nsd_study_outcome', E::ts('Study outcome'), [
       'rows' => 4,
       'cols' => 100,
       'disabled' => 'disabled',
@@ -216,6 +229,7 @@ class CRM_Nihrbackbone_Form_NbrStudy extends CRM_Core_Form {
       // status pending when adding
       case CRM_Core_Action::ADD:
         $defaults['status_id'] = CRM_Nihrbackbone_BackboneConfig::singleton()->getPendingStudyStatus();
+        $defaults['nsd_prevent_upload_portal'] = TRUE;
         break;
       case CRM_Core_Action::UPDATE:
       case CRM_Core_Action::VIEW:
@@ -229,6 +243,7 @@ class CRM_Nihrbackbone_Form_NbrStudy extends CRM_Core_Form {
           }
           $defaults[$key] = $value;
         }
+        $defaults['researcher_id'] = implode(",", CRM_Nihrbackbone_BAO_NbrStudyResearcher::getStudyResearchers($this->_studyId));
         break;
     }
     return $defaults;
@@ -481,6 +496,11 @@ class CRM_Nihrbackbone_Form_NbrStudy extends CRM_Core_Form {
     try {
       $createdCampaign = civicrm_api3('Campaign', 'create', $this->setCampaignParameters());
       $this->_studyId = $createdCampaign['id'];
+      $researcherIds = [];
+      if (isset($this->_submitValues['researcher_id'])) {
+        $researcherIds = explode(",", $this->_submitValues['researcher_id']);
+      }
+      CRM_Nihrbackbone_BAO_NbrStudyResearcher::saveStudyResearchers($this->_studyId, $researcherIds, $this->_action);
       $studyNumber = CRM_Nihrbackbone_NbrStudy::getStudyNumberWithId($this->_studyId);
       CRM_Core_Session::setStatus(E::ts("Saved study ") . $studyNumber, E::ts('Saved'), 'success');
     }