+++ CRM/Nihrbackbone/Upgrader.php
@@ -469,6 +469,62 @@ class CRM_Nihrbackbone_Upgrader extends CRM_Nihrbackbone_Upgrader_Base {
   }
 
   /**
+   * Upgrade 1160 - add study outcome field (see https://www.wrike.com/open.htm?id=1013253659)
+   *
+   * @return bool
+   * @throws
+   */
+  public function upgrade_1160() {
+    $this->ctx->log->info(E::ts('Applying update 1160 - add study outcome to study data'));
+    // only if custom group for study data is present
+    if (CRM_Core_DAO::checkTableExists("civicrm_value_nbr_study_data")) {
+      // add custom field scientific info
+      if (!CRM_Core_BAO_SchemaHandler::checkIfFieldExists("civicrm_value_nbr_study_data", "nsd_study_outcome")) {
+        \Civi\Api4\CustomField::create()
+          ->addValue('custom_group_id:name', 'nbr_study_data')
+          ->addValue('label', 'Study Outcome')
+          ->addValue('html_type', 'RichTextEditor')
+          ->addValue('data_type', 'Memo')
+          ->addValue('is_active', TRUE)
+          ->addValue('is_searchable', FALSE)
+          ->addValue('in_selector', TRUE)
+          ->addValue('name', 'nsd_study_outcome')
+          ->addValue('column_name', 'nsd_study_outcome')
+          ->execute();
+      }
+    }
+    return TRUE;
+  }
+
+  /**
+   * Upgrade 1170 - add entity NbrStudyResearcher table (see https://www.wrike.com/open.htm?id=933254901)
+   *              - migrate existing study researchers
+   *
+   * @return bool
+   * @throws
+   */
+  public function upgrade_1170(): bool {
+    $this->ctx->log->info(E::ts('Applying update 1170 - add table for entity NbrStudyResearcher'));
+    // create new table for entity
+    if (!CRM_Core_DAO::checkTableExists('civicrm_nbr_study_researcher')) {
+      $create = "CREATE TABLE `civicrm_nbr_study_researcher` (`id` int unsigned NOT NULL AUTO_INCREMENT COMMENT 'Unique NbrStudyResearcher ID',
+         `researcher_contact_id` int unsigned COMMENT 'FK to Contact', `nbr_study_id` int unsigned COMMENT 'FK to Campaign', PRIMARY KEY (`id`),
+         CONSTRAINT FK_civicrm_nbr_study_researcher_researcher_contact_id FOREIGN KEY (`researcher_contact_id`) REFERENCES `civicrm_contact`(`id`) ON DELETE CASCADE,
+         CONSTRAINT FK_civicrm_nbr_study_researcher_nbr_study_id FOREIGN KEY (`nbr_study_id`) REFERENCES `civicrm_campaign`(`id`) ON DELETE CASCADE)
+    ENGINE=InnoDB;";
+      CRM_Core_DAO::executeQuery($create);
+      // migrate existing researchers
+      $dao = CRM_Core_DAO::executeQuery("SELECT nsd_researcher, entity_id AS study_id FROM civicrm_value_nbr_study_data WHERE nsd_researcher IS NOT NULL");
+      while ($dao->fetch()) {
+        if (!empty($dao->study_id) && !empty($dao->nsd_researcher)) {
+          CRM_Nihrbackbone_BAO_NbrStudyResearcher::createStudyResearcher((int) $dao->study_id, (int) $dao->nsd_researcher);
+        }
+      }
+    }
+    return TRUE;
+  }
+
+  /**
    * Swap values for unable to willing columns
    *
    * @param $columns